Get rid of docs/make_local.jl
